+++ utils/plot_def.py
@@ -9,7 +9,7 @@ import matplotlib.pyplot as plt
 color = ['#1f77b4', '#ff7f0e', '#2ca02c', '#d62728', '#9467bd', '#8c564b', '#e377c2', '#7f7f7f', '#bcbd22', '#17becf']
 
 # Plot parameters and sizes
-plt.rc('font', size = 16)
+plt.rc('font', size = 14)
 plt.rc('axes', linewidth=1, titlesize=14)     # fontsize of the axes title
 plt.rc('axes', labelsize=14)    # fontsize of the x and y labels
 plt.rc('xtick', top=True, bottom=True, direction='in')