+++ utils/plot_functions_UI.py
@@ -5,7 +5,7 @@ import matplotlib.pyplot as plt
 import streamlit as st
 from pySIMsalabim.plots import plot_functions as utils_plot
 
-######### Function Definitions ####################################################################    
+######### Function Definitions ####################################################################   
 
 def plot_result_JV(data, choice_voltage, plot_funcs, ax, exp, data_exp='', xscale = 'linear', yscale = 'linear'):
     """Make a plot of the JV curve based on the 'JV.dat" file
@@ -35,6 +35,12 @@ def plot_result_JV(data, choice_voltage, plot_funcs, ax, exp, data_exp='', xscal
         Updated Axes object for the plot
     """
 
+    # Remove the +/- toggles on number inputs
+    st.markdown("""<style>
+        button.step-up {display: none;}
+        button.step-down {display: none;}
+        div[data-baseweb] {border-radius: 4px;}
+        </style>""",unsafe_allow_html=True) 
 
     # If the yscale is log, we need to take the absolute value of the current to not lose the negative part of the curve. 
     if yscale == 'log':
@@ -105,7 +111,7 @@ def get_nonzero_parameters(pars, data, choice_voltage):
 
 def create_UI_component_plot(data_org, pars, x_key, xlabel, ylabel, title, plot_no, fig, ax, plot_type,
                              cols, choice_voltage = 0, source_type = '', show_plot_param=True, show_yscale=True, yscale_init=0, xscale_init=0, 
-                             show_xscale=False, weight_key = '', weight_label = '', weight_norm = 'linear', error_x = '', error_y='', show_legend=True,error_fmt='-'):
+                             show_xscale=False, show_xrange = True, show_yrange = True, xrange_format = "%f", yrange_format = "%e", xrange_val = None, yrange_val = None, weight_key = '', weight_label = '', weight_norm = 'linear', error_x = '', error_y='', show_legend=True,error_fmt='-'):
     """Create a plot for the provided data and place it into a column structure. 
     Add the plot options to the right of the plot when needed. 
     When plotting a 'Var' type file, plot only for the selected voltage
@@ -148,6 +154,18 @@ def create_UI_component_plot(data_org, pars, x_key, xlabel, ylabel, title, plot_
         Initial scale of the x-axis, used in xscale_options, by default 0
     show_xscale : bool, optional
         Show a radio toggle to switch between a linear or log x scale, by default true
+    show_xrange : bool, optional
+        Show input fields to set the x-range, by default True
+    show_yrange : bool, optional
+        Show input fields to set the y-range, by default True
+    xrange_format : string, optional
+        Format of the x-range input fields, by default "%f"
+    yrange_format : string, optional
+        Format of the y-range input fields, by default "%e"
+    xrange_val : List, optional
+        Initial values for the x-range input fields, by default [None, None] in which case the min,max values of the data are used as limits
+    yrange_val : List, optional
+        Initial values for the y-range input fields, by default [None, None] in which case the min,max values of the data are used as limits
     weight_key : string, optional
         Key in the dataframe for the colorbar data, ignored if empty string, by default ''
     weight_label : string, optional
@@ -170,6 +188,13 @@ def create_UI_component_plot(data_org, pars, x_key, xlabel, ylabel, title, plot_
     Axes
         Updated Axes object 
     """
+
+    # Remove the +/- toggles on number inputs
+    st.markdown("""<style>
+        button.step-up {display: none;}
+        button.step-down {display: none;}
+        div[data-baseweb] {border-radius: 4px;}
+        </style>""",unsafe_allow_html=True) 
     
     scale_options = ['linear', 'log']
     error_options = [True, False]
@@ -187,7 +212,7 @@ def create_UI_component_plot(data_org, pars, x_key, xlabel, ylabel, title, plot_
             st.markdown('<br>', unsafe_allow_html=True)
             st.markdown('<hr>', unsafe_allow_html=True)
 
-        with st.expander('Figure options', expanded=False):
+        with st.expander('Figure options', expanded=True):
             # Select which parameters to plot
             if show_plot_param:
                 options = st.multiselect('Parameters to plot:', list(pars.keys()), list(pars.keys()), key = str(plot_no) + '-par-options')
@@ -198,15 +223,76 @@ def create_UI_component_plot(data_org, pars, x_key, xlabel, ylabel, title, plot_
 
             # Select the y-scale
             if show_yscale:
-                yscale = st.radio('y-scale', scale_options, index = yscale_init, key = str(plot_no) + '-y-scale')
+                st.text('y-scale:')
+                yscale = st.radio('y-scale', scale_options, index = yscale_init, key = str(plot_no) + '-y-scale', label_visibility='collapsed',horizontal=True)
             else:
                 yscale = scale_options[yscale_init]
 
             # Select the x-scale
             if show_xscale:
-                xscale = st.radio('x-scale', scale_options, index = xscale_init, key = str(plot_no) + '-x-scale')
+                st.text('x-scale:')
+                xscale = st.radio('x-scale', scale_options, index = xscale_init, key = str(plot_no) + '-x-scale', label_visibility='collapsed',horizontal=True)
             else:
                 xscale = scale_options[xscale_init]
+
+            # Show the x,y range input fields to rescale the plot
+            if show_xrange:
+                if xrange_val is None:
+                    # Use the min and max values of the selected parameter as limits with a margin of 5% of the range to prevent the data from merging with the axis
+                    xlow_init = data[x_key].min()
+                    xup_init = data[x_key].max()
+
+                    # Use 5% of the interval
+                    x_5p = abs((xup_init - xlow_init) * 0.05)
+                    # Subtract this value from the lower limit
+                    xlow_init -= x_5p
+                    # Add this value to the upper limit
+                    xup_init += x_5p
+
+                    # round to two decimals to not display the full float
+                    if 'f' in xrange_format:
+                        xlow_init = round(xlow_init,2)
+                        xup_init = round(xup_init,2)
+                else:
+                    # Limit values provided by the user as arguments
+                    xlow_init = xrange_val[0]
+                    xup_init = xrange_val[1]
+  
+                st.text('Set the x-axis range:')
+                col1_x, col2_x, = st.columns([1,1])
+                with col1_x:
+                    xlow = st.number_input('X-range', value = xlow_init, key = str(plot_no) + '-xrange_low',label_visibility="collapsed", format=xrange_format)
+                with col2_x:
+                    xup = st.number_input('X-range', value=xup_init, key = str(plot_no) + '-xrange_up',label_visibility="collapsed", format=xrange_format)
+
+            if show_yrange:
+                if yrange_val is None:
+                    # Get the lowest and highest value of the selected parameters
+                    ylow_init = data[options[0]].min()
+                    yup_init = data[options[0]].max()
+                    for par in options:
+                        ylow_init = min(ylow_init, data[par].min())
+                        yup_init = max(yup_init, data[par].max())
+
+                    # Use 5% of the interval
+                    y_5p = abs((yup_init - ylow_init) * 0.05)
+
+                    # In a log plot, the lower limit should be above zero
+                    if yscale == 'log' and ylow_init - y_5p <= 0:
+                        ylow_init = ylow_init
+                    else:
+                        # Subtract this value from the lower limit
+                        ylow_init -= y_5p
+                    # Add this value to the upper limit
+                    yup_init += y_5p
+
+                st.text('Set the y-axis range:')
+                col1_y, col2_y, = st.columns([1,1])
+  
+                with col1_y:
+                    ylow = st.number_input('Y-range', value=ylow_init, key = str(plot_no) + '-yrange_low',label_visibility="collapsed", format=yrange_format)
+                with col2_y:
+                    yup = st.number_input('Y-range', value=yup_init, key = str(plot_no) + '-yrange_up',label_visibility="collapsed", format=yrange_format)
 
             # Show the error margin on y1 axis
             if plot_type == plt.errorbar:
@@ -238,6 +324,12 @@ def create_UI_component_plot(data_org, pars, x_key, xlabel, ylabel, title, plot_
             else:
                 ax = utils_plot.plot_result(data, pars, options, x_key, xlabel, ylabel,xscale, yscale, title, ax, plt.plot,legend=show_legend)
         
+        # Set the x,y range, independent of plot type and/or errorbars
+        if show_xrange:
+            ax.set_xlim(xlow, xup)
+        if show_yrange:
+            ax.set_ylim(ylow, yup)
+
         return fig,ax
 
 def create_UI_component_plot_twinx(data_org, pars, selected_1, selected_2, x_key, xlabel, ylabel_1, ylabel_2, title, fig, ax_1, ax_2, 
@@ -301,7 +393,7 @@ def create_UI_component_plot_twinx(data_org, pars, selected_1, selected_2, x_key
             st.markdown('<br>', unsafe_allow_html=True)
             st.markdown('<hr>', unsafe_allow_html=True)
 
-        with st.expander('Figure options', expanded=False):
+        with st.expander('Figure options', expanded=True):
             # Select which parameters to plot
             if show_plot_param:
                 options = st.multiselect('Parameters to plot:', list(pars.keys()), list(pars.keys()), key = str(1) + '-par-options' + str(selected_1))
@@ -310,23 +402,95 @@ def create_UI_component_plot_twinx(data_org, pars, selected_1, selected_2, x_key
             else:
                 options =  list(pars.keys())
 
+            # Show the error margin on y1 axis
+            if show_errors:
+                st.text('Show error margins::')
+                yerror = st.radio('Show error margins', error_options, index = 0, key = str(1) + '-y-error' + str(selected_1), label_visibility='collapsed',horizontal=True)
+            else:
+                yerror = error_options[1]
+
             # Select the y1-scale
             if show_yscale_1:
-                yscale_1 = st.radio('y-scale (Left)', scale_options, index = yscale_init_1, key = str(1) + '-y1-scale' + str(selected_1))
+                st.text('y-scale (Left):')
+                yscale_1 = st.radio('y-scale (Left)', scale_options, index = yscale_init_1, key = str(1) + '-y1-scale' + str(selected_1), label_visibility='collapsed',horizontal=True)
             else:
                 yscale_1 = scale_options[yscale_init_1]
 
             # Select the y2-scale
             if show_yscale_2:
-                yscale_2 = st.radio('y-scale (Right)', scale_options, index = yscale_init_2, key = str(1) + '-y2-scale' + str(selected_1))
+                st.text('y-scale (Right):')
+                yscale_2 = st.radio('y-scale (Right)', scale_options, index = yscale_init_2, key = str(1) + '-y2-scale' + str(selected_1), label_visibility='collapsed',horizontal=True)
             else:
                 yscale_2 = scale_options[yscale_init_2]
 
-            # Show the error margin on y1 axis
-            if show_errors:
-                yerror = st.radio('Show error margins', error_options, index = 0, key = str(1) + '-y-error' + str(selected_1))
+            # Set the x range
+            # Use the min and max values of the selected parameter as limits with a margin of 5% of the range to prevent the data from merging with the axis
+            xlow_init = data_org[x_key].min()
+            xup_init = data_org[x_key].max()
+
+            # Use 5% of the interval
+            x_5p = abs((xup_init - xlow_init) * 0.05)
+            # Subtract this value from the lower limit
+            xlow_init -= x_5p
+            # Add this value to the upper limit
+            xup_init += x_5p
+
+            st.text('Set the x-axis range:')
+            col1_x, col2_x, = st.columns([1,1])
+            with col1_x:
+                xlow = st.number_input('X-range', value = xlow_init, key = 'Bode-xrange_low-' + str(selected_1),label_visibility="collapsed", format="%.2e")
+            with col2_x:
+                xup = st.number_input('X-range', value=xup_init, key = 'Bode-xrange_up-' + str(selected_1),label_visibility="collapsed", format="%.2e")
+
+            # Set the y (Left) range
+            # Get the lowest and highest value of the selected parameters
+            ylow_init_left = data_org[selected_1].min()[0]
+            yup_init_left = data_org[selected_1].max()[0]
+
+            # Use 5% of the interval
+            y_5p_left = abs((yup_init_left - ylow_init_left) * 0.05)
+
+            # In a log plot, the lower limit should be above zero
+            if yscale_1 == 'log' and ylow_init_left - y_5p_left <= 0:
+                ylow_init_left = ylow_init_left
             else:
-                yerror = error_options[1]
+                # Subtract this value from the lower limit
+                ylow_init_left -= y_5p_left
+            # Add this value to the upper limit
+                yup_init_left += y_5p_left
+
+            st.text('Set the y-axis range (Left):')
+            col1_y_left, col2_y_left, = st.columns([1,1])
+
+            with col1_y_left:
+                ylow_left = st.number_input('Y-range', value=ylow_init_left, key = 'Bode-left-yrange_low-' + str(selected_1),label_visibility="collapsed", format="%.2e")
+            with col2_y_left:
+                yup_left = st.number_input('Y-range', value=yup_init_left, key = 'Bode-left-yrange_up-' + str(selected_1),label_visibility="collapsed", format="%.2e")
+
+            # Set the y (Right) range
+            # Get the lowest and highest value of the selected parameters
+            ylow_init_right = data_org[selected_2].min()[0]
+            yup_init_right = data_org[selected_2].max()[0]
+
+            # Use 5% of the interval
+            y_5p_right = abs((yup_init_right - ylow_init_right) * 0.05)
+
+            # In a log plot, the lower limit should be above zero
+            if yscale_2 == 'log' and ylow_init_right - y_5p_right <= 0:
+                ylow_init_right = ylow_init_right
+            else:
+                # Subtract this value from the lower limit
+                ylow_init_right -= y_5p_right
+            # Add this value to the upper limit
+                yup_init_right += y_5p_right
+
+            st.text('Set the y-axis range (Right):')
+            col1_y_right, col2_y_right, = st.columns([1,1])
+
+            with col1_y_right:
+                ylow_right = st.number_input('Y-range', value=ylow_init_right, key = 'Bode-right-yrange_low-' + str(selected_2),label_visibility="collapsed", format="%.2e")
+            with col2_y_right:
+                yup_right = st.number_input('Y-range', value=yup_init_right, key = 'Bode-right-yrange_up-' + str(selected_2),label_visibility="collapsed", format="%.2e")
 
     with cols[1]:
         # Create plot
@@ -336,4 +500,9 @@ def create_UI_component_plot_twinx(data_org, pars, selected_1, selected_2, x_key
         else:
             ax = utils_plot.plot_result_twinx(data_org, pars, selected_1, selected_2, x_key, xlabel, ylabel_1, ylabel_2, 'log', yscale_1, yscale_2, title,ax_1,ax_2, 
                                         plt.plot)
+            
+        ax.set_xlim(xlow, xup)
+        ax_1.set_ylim(ylow_left, yup_left)
+        ax_2.set_ylim(ylow_right, yup_right)
+            
         st.pyplot(fig, format='png')